Update bin/jupyter-kernel-gap
To find GAP, we now proceed as follows: First, if $JUPYTER_GAP_EXECUTABLE is
set, we use that. If not, then we check if $GAP is set and use that. If not,
we try to use `gap` in the current path.

Also avoid the use of `which`, which is non-standard. Instead, just hard code
the name `gap`, and don't bother performing any further checks (we don't
validate the value of $JUPYTER_GAP_EXECUTABLE or $GAP either)

Finally, use `exec` to start GAP, replacing the shell, instead of creating
a fresh process.

The script now passes the tests on https://www.shellcheck.net