Use GAP library functions to handle errors
Now that the GAP error functions can redirect their output to a
chosen stream, the previously used hack is no longer needed.

This commit provides a function JupyterOnBreak which prints a
traceback and then skips the break loop by calling JUMP_TO_CATCH.
When the jupyter kernel runs OnBreak is overwritten with
JupyterOnBreak.

To use this GAP is started without the -T option.

Fixes #38.